% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fits_MM.R
\docType{data}
\name{all_fits_MM}
\alias{all_fits_MM}
\title{A list of fitted models under Markov assumption}
\format{
A \code{list} of 10 elements, where each element is a \code{list} of 3 elements. Each sublist contains
the parameter estimates (estimands) for a specific transition, obtained from the respective fitted model.
}
\source{
Estimands generated by fitting parametric models over datasets that were generated using multiple imputation.
}
\usage{
all_fits_MM
}
\description{
A list of fitted models under Markov assumption
}
\keyword{datasets}
