% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averaging_params.R
\name{averaging_params}
\alias{averaging_params}
\title{Averaging parameter estimates across all fitted models for each transition}
\usage{
averaging_params(all_fits)
}
\arguments{
\item{all_fits}{A \code{list} of previously fitted models, where each element in the list
is itself a list containing the model coefficients for each transition.}
}
\value{
A \code{data.frame} with rows representing transitions and columns representing
parameters. The values in the \code{data.frame} are the averaged parameter estimates
across different models.
}
\description{
This function averages the parameter estimates obtained from multiple fitted models
across each transition. It returns a \code{data.frame} where each row corresponds to
a transition and each column represents a parameter, containing the mean value
of each parameter across all provided models.
}
\examples{
avg_params <- averaging_params(all_fits_MM)
print(avg_params)
}
