% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_parameters_MM.R
\docType{data}
\name{avg_parameters_MM}
\alias{avg_parameters_MM}
\title{A \code{matrix} containing the mean estimates of model parameters.}
\format{
A\code{matrix} where each row corresponds to the parameter estimates for a specific transition.
}
\source{
Estimates derived by averaging parameter estimates obtained from fitting parametric models
to datasets generated through multiple imputation.
}
\usage{
avg_parameters_MM
}
\description{
A \code{matrix} containing the mean estimates of model parameters.
}
\keyword{datasets}
