% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fits a parametric illness-death model with known disease progression}
\usage{
fit_model(data, cov_vector, clock_assumption, distribution)
}
\arguments{
\item{data}{A \code{data.frame} containing imputed disease onset times, disease status, and covariates for each individual.}

\item{cov_vector}{vector with covariates names contained in data.}

\item{clock_assumption}{A character string specifying the assumed time scale:
\itemize{
\item \code{"forward"} (Markov assumption): Only the time at transition is considered.
\item \code{"mix"} (Semi-Markov assumption): Accounts for the time spent in the disease-free state.
}}

\item{distribution}{A character string specifying the parametric form of baseline hazards.
Must be one of the distributions available in \code{flexsurv::flexsurvreg}, e.g., \code{"weibull"}, \code{"exponential"}, \code{"gompertz"}.}
}
\value{
A \code{list} of fitted \code{flexsurvreg} models, one per transition in the illness-death model.
}
\description{
This function fits a parametric illness-death model leveraging the \code{mstate} and \code{flexsurv} packages,
assuming either a Markov or Semi-Markov structure for disease progression.
}
\examples{
# Example usage:
data <- imputed_dataset_MM # Example dataset with required variables
fit <- fit_model(data,c("cov1","cov2","cov3"), clock_assumption = "forward", distribution = "weibull")
summary(fit[[1]]) # Summary of first transition model
}
