% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ic_computation.R
\name{Ic_computation}
\alias{Ic_computation}
\title{Compute Confidence Intervals using Rubin's Rule}
\usage{
Ic_computation(avg_parameters, all_fits)
}
\arguments{
\item{avg_parameters}{A \code{matrix} containing the average parameter estimates across imputations.}

\item{all_fits}{A \code{list} of fitted models from multiple imputed datasets.}
}
\value{
A \code{list} with the same number of elements as the transitions of the multi-state model,
each containing a matrix with the lower and upper 95\% confidence interval bounds
for the parameters associated with a specific transition.
}
\description{
This function calculates confidence intervals for model parameters using Rubin's rule
for multiple imputations to take into account within and between variability.
}
\examples{
IC <- Ic_computation(avg_parameters_MM, all_fits_MM)
print(IC)

}
