% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dataset.R
\name{prepare_dataset}
\alias{prepare_dataset}
\title{Prepare dataset to perform multiple imputation of the disease history}
\usage{
prepare_dataset(data)
}
\arguments{
\item{data}{A \code{data.table} or \code{data.frame} that must contain the following columns:
\itemize{
\item \code{patient_id}: Unique identifier for each patient (numeric).
\item \code{dead}: Binary indicator (0/1) for whether the patient is dead.
\item \code{death_time}: Time of death if it has occured or censoring time otherwise (numeric).
\item \code{onset}: Binary indicator (0/1) for disease onset.
\item \code{onset_age}: Age at disease onset if it has occured or death_time otherwise (numeric).
\item \code{age}: Patient's current age at that specific visit (numeric).
\item \code{visits}: Indicator of the current visit (numeric).
The \code{data.frame} could contain extra columns with covariates values.
}}
}
\value{
A \code{list} containing two \code{data.frame}s:
\itemize{
\item The first \code{data.frame} retains all columns from the input data while adding \code{last_bfo}, which records
the patient's age at their last visit before diagnosis (\code{onset == 1}). It includes only the first row
per patient.
\item The second \code{data.frame} is an ordered copy of the input data.
}

Both \code{data.frame}s have updated, sequentially ordered \code{patient_id} values.
}
\description{
Prepare dataset to perform multiple imputation of the disease history
}
\examples{
data(toy_data_MM)
prepare_dataset(toy_data_MM)
}
