% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mipd.R
\name{run_mipd}
\alias{run_mipd}
\title{Multiple imputation for interval-censored illness-death model and fitting Parametric Illness-Death Model over imputed datasets}
\usage{
run_mipd(
  data,
  m,
  clock_assumption,
  cov_vector,
  distribution,
  inner_cores,
  max_iter,
  eps,
  boot = FALSE
)
}
\arguments{
\item{data}{A \code{data.table} or \code{data.frame} containing the following columns:
\itemize{
\item \code{patient_id}: Unique identifier for each patient (numeric).
\item \code{dead}: Binary indicator (0/1) of whether the patient is dead.
\item \code{death_time}: Time of death if it has occurred, or censoring time otherwise (numeric).
\item \code{onset}: Binary indicator (0/1) of disease onset.
\item \code{onset_age}: Age at disease onset if it has occurred, or death time otherwise (numeric).
\item \code{age}: Patient's current age at the specific visit (numeric).
\item \code{visits}: Indicator of the current visit (numeric).
Additional columns in the \code{data.frame} will be treated as covariates.
}}

\item{m}{A numeric value specifying the number of imputations to perform.
We recommend choosing a value between 20 and 50 to ensure consistency.
Keep in mind that the computational time increases with the number of imputations (\code{m}),
so for large datasets, it may be more efficient to choose a smaller value like \code{m = 30}.}

\item{clock_assumption}{A character string specifying the assumed time scale:
\itemize{
\item \code{"forward"} (Markov assumption): Only the time at transition is considered.
\item \code{"mix"} (Semi-Markov assumption): Accounts for the time spent in the disease-free state.
}}

\item{cov_vector}{vector of covariates names.}

\item{distribution}{A character string specifying the parametric form of the baseline hazard.
It must be one of the distributions available in \code{flexsurv::flexsurvreg}.}

\item{max_iter}{maximum number of iterations.}

\item{eps}{tolerance for convergence.}
}
\value{
A \code{list} containing the average values of model parameter estimates in the first element
and 95\% confidence intervals computed using Rubin's rule saved by transition in the second element.
}
\description{
Performs multiple imputation on a dataset containing interval-censored data
about the disease progression of individuals, and fits a parametric illness-death model
assuming exact observation times over each dataset with imputed disease status and disease onset.
}
\examples{
# Example usage:
fit <- run_mipd(
toy_data_MM,
m = 5,
cov_vector = c("cov1", "cov2", "cov3"),
clock_assumption = "forward",
distribution = "gompertz",
inner_cores = 1,
max_iter = 20,
eps = 1e-3
)
)
print(fit)
}
